client: added common header for the issue and project
