+++ client/src/app/issue/issue-list/index.issue.list.js
@@ -1,29 +1,13 @@
 import React, { Fragment, useEffect } from "react";
-import Link from "react-router-dom/Link";
 import { connect } from "react-redux";
-import withStyles from "@material-ui/core/styles/withStyles";
-import Typography from "@material-ui/core/Typography";
 import Grid from "@material-ui/core/Grid";
-import Button from "@material-ui/core/Button";
-
-import AddIcon from "@material-ui/icons/Add";
 
 import IssueList from "../../../components/issue/issue.list";
+import Header from "../../../components/header/header";
 import Loader from "../../../components/loader/loader.page";
 import { issueList } from "../../../store/actions/index.action";
 
-const styles = ({ spacing }) => ({
-    addIcon: {
-        marginRight: spacing.unit / 2.5
-    },
-    headerMargin: {
-        marginBottom: spacing.unit * 2
-    }
-});
-
-const _Link = props => <Link to="/d/issues/add" {...props} />;
-
-const IssueListIndex = ({ classes, $issueList, _issueList }) => {
+const IssueListIndex = ({ $issueList, _issueList }) => {
     useEffect(() => {
         $issueList();
     }, []);
@@ -34,27 +18,7 @@ const IssueListIndex = ({ classes, $issueList, _issueList }) => {
 
     return (
         <Fragment>
-            <Grid
-                container
-                justify="space-between"
-                alignItems="center"
-                className={classes.headerMargin}
-            >
-                <Grid item>
-                    <Typography variant="h5">Issues</Typography>
-                </Grid>
-                <Grid item>
-                    <Button
-                        variant="contained"
-                        color="primary"
-                        size="small"
-                        component={_Link}
-                    >
-                        <AddIcon className={classes.addIcon} />
-                        Add
-                    </Button>
-                </Grid>
-            </Grid>
+            <Header title="Issues" addLink="/d/issues/add" />
             <Grid container>
                 <Grid item xs={12}>
                     <IssueList items={_issueList} />
@@ -75,4 +39,4 @@ const mapDispatchToProps = dispatchEvent => ({
 export default connect(
     mapStateToProps,
     mapDispatchToProps
-)(withStyles(styles)(IssueListIndex));
+)(IssueListIndex);
